% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltmleMediation_TMLECalcs.R
\name{CalcCumG}
\alias{CalcCumG}
\title{CalcCumG}
\usage{
CalcCumG(g, gbounds)
}
\arguments{
\item{g}{Estimate of \code{g} derived from \code{EstimateG()}.}

\item{gbounds}{Lower and upper bounds for g.}
}
\value{
Returns cummulative G for multiple time points.
}
\description{
Calculate bounded cumulative G.
}
