% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltmleMediation_TMLECalcs.R
\name{CalcG}
\alias{CalcG}
\title{CalcG}
\usage{
CalcG(prob.A.is.1, cur.abar, deterministic.newdata)
}
\arguments{
\item{prob.A.is.1}{TO DO}

\item{cur.abar}{TO DO}

\item{deterministic.newdata}{TO DO}
}
\value{
Returns G.
}
\description{
Calculate G
}
