% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltmleMediation_TMLECalcs.R
\name{CalcGUnboundedToBoundedRatio}
\alias{CalcGUnboundedToBoundedRatio}
\title{CalcGUnboundedToBoundedRatio}
\usage{
CalcGUnboundedToBoundedRatio(g.list, nodes, final.Ynodes)
}
\arguments{
\item{g.list}{TO DO}

\item{nodes}{All available nodes in the data.}

\item{final.Ynodes}{Final Y nodes.}
}
\value{
Returns...
}
\description{
In case the GLM did not converge and the TMLE does not solve the score equation, attempt to solve the score equation directly.
}
