% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltmleMediation_TMLECalcs.R
\name{CalcIC}
\alias{CalcIC}
\title{CalcIC}
\usage{
CalcIC(
  Qstar.kplus1,
  Qstar,
  h.g.ratio,
  uncensored,
  intervention.match,
  regimes.with.positive.weight
)
}
\arguments{
\item{Qstar.kplus1}{Estimate of the expectation with respect to the distribution of the node one ahead of the current node given the past (dimension: n x num.regimes).}

\item{Qstar}{Possibly targeted estimate of Q for the current node.}

\item{h.g.ratio}{I() * h * observation.weights * cum.q.ratio / g.}

\item{uncensored}{Uncensored samples}

\item{intervention.match}{Samples that match the rule.}

\item{regimes.with.positive.weight}{Regimes with positive weight.}
}
\value{
Returns IC for one node Q estimate.
}
\description{
Calculate the TMLE influence curve for one node Q estimate.
}
