% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medltmle_HelperFun.R
\name{CalcInterventionMatchArray}
\alias{CalcInterventionMatchArray}
\title{CalcInterventionMatchArray}
\usage{
CalcInterventionMatchArray(data, regimes, Anodes)
}
\arguments{
\item{data}{TO DO}

\item{regimes}{TO DO}

\item{Anodes}{TO DO}
}
\value{
Returns intervention that matches the array.
}
\description{
Calculates a logical array - n x num.Anodes x num.regimes = follows regime[j] up to and including Anode[i]
}
