% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltmleMediation_TMLECalcs.R
\name{CalcIPTWMediation}
\alias{CalcIPTWMediation}
\title{CalcIPTWMediation}
\usage{
CalcIPTWMediation(inputs, cum.g.abar, cum.qz.abar, cum.qz.abar.prime,
  msm.weights)
}
\arguments{
\item{inputs}{Output of \code{CreateMedInputs}}

\item{cum.g.abar}{TO DO}

\item{cum.qz.abar}{TO DO}

\item{cum.qz.abar.prime}{TO DO}

\item{msm.weights}{}
}
\value{
Returns estimate of L densities.
}
\description{
If not is.Z.dens, estimating L densities
}
