% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medltmle_HelperFun.R
\name{CalcUncensoredMatrix}
\alias{CalcUncensoredMatrix}
\title{CalcUncensoredMatrix}
\usage{
CalcUncensoredMatrix(data, Cnodes)
}
\arguments{
\item{data}{TO DO}

\item{Cnodes}{TO DO}
}
\value{
Calc logical matrix - n x numCnodes = is uncensored up to and including Cnode[i]
}
\description{
Determine which patients have died or have Q set deterministically by user function before cur.node.
}
