% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medltmle_HelperFun.R
\name{CleanData}
\alias{CleanData}
\title{CleanData}
\usage{
CleanData(data, nodes, deterministic.Q.function, survivalOutcome,
  showMessage = TRUE)
}
\arguments{
\item{data}{TO DO}

\item{nodes}{TO DO}

\item{deterministic.Q.function}{TO DO}

\item{survivalOutcome}{TO DO}

\item{showMessage}{TO DO}
}
\value{
Returns cleaned data.
}
\description{
Set all nodes (except Y) to NA after death or censoring; Set Y nodes to 1 after death.
}
