% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medltmle_HelperFun.R
\name{ConvertCensoringNodes}
\alias{ConvertCensoringNodes}
\title{ConvertCensoringNodes}
\usage{
ConvertCensoringNodes(data, Cnodes, has.deterministic.functions = FALSE)
}
\arguments{
\item{data}{TO DO}

\item{Cnodes}{TO DO}

\item{has.deterministic.functions}{TO DO}
}
\value{
Returns data with converted censoring nodes.
}
\description{
Convert censoring nodes stored as binaries into factors (factors are recommended but binaries are currently accepted)
}
