% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medltmle_HelperFun.R
\name{ConvertCensoringNodesToBinary}
\alias{ConvertCensoringNodesToBinary}
\title{ConvertCensoringNodesToBinary}
\usage{
ConvertCensoringNodesToBinary(data, Cnodes)
}
\arguments{
\item{data}{TO DO}

\item{Cnodes}{TO DO}
}
\value{
Returns data with censoring nodes converted to binary.
}
\description{
Before passing data to SuperLearner, convert factors to binary.
}
