% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medltmle_HelperFun.R
\name{ConvertToMainTerms}
\alias{ConvertToMainTerms}
\title{ConvertToMainTerms}
\usage{
ConvertToMainTerms(data, msm, summary.measures, nodes)
}
\arguments{
\item{data}{TO DO}

\item{msm}{TO DO}

\item{summary.measures}{TO DO}

\item{nodes}{}
}
\value{
Returns main terms Marginal Structural Model, summary measures, beta names and baseline column names.
}
\description{
Converts a general formula to a main terms formula and combine summary measures with baseline covariates.
Ex: If working.msm is "Y ~ X1*X2", convert to "Y ~ -1 + S1 + S1 + S3 + S4" where S1 is 1 (intercept), S2 is X1, S3 is X2, S4 is X1:X2.
}
