% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medltmle_CreateMedInputs.R
\name{CreateMedInputs}
\alias{CreateMedInputs}
\title{CreateMedInputs}
\usage{
CreateMedInputs(data, Anodes, Cnodes, Lnodes, Ynodes, Znodes, survivalOutcome,
  QLform, QZform, gform, qzform, qLform, gbounds, Yrange,
  deterministic.g.function, SL.library, regimes, regimes.prime, working.msm,
  summary.measures, final.Ynodes, stratify, msm.weights, estimate.time, gcomp,
  iptw.only, deterministic.Q.function, IC.variance.only, observation.weights)
}
\arguments{
\item{data}{Dataframe containing the data in a wide format.}

\item{Anodes}{names of columns containing A covariates.}

\item{Cnodes}{names of columns containing C covariates.}

\item{Lnodes}{names of columns containing L covariates.}

\item{Ynodes}{names of columns containing Y covariates.}

\item{Znodes}{names of columns containing Z covariates.}

\item{survivalOutcome}{logical variable specifying if the outcome is survival.}

\item{QLform}{Q forms for L covariates.}

\item{QZform}{Q forms for Z covariates.}

\item{gform}{g form for intervention (if there is a censoring variable, include C as well).}

\item{qzform}{g form for Z covariates.}

\item{qLform}{g form for L covariates.}

\item{gbounds}{Bounds for the propensity score.}

\item{Yrange}{Specify range for the outcome.}

\item{deterministic.g.function}{Logical specifying if g is a deterministic function.}

\item{SL.library}{SuperLearner library for estimation.}

\item{regimes}{Regimes for abar as output by the RegimesFromAbar function.}

\item{regimes.prime}{Regimes for abar.prime as output by the RegimesFromAbar function.}

\item{working.msm}{Working MSM as output by the GetMediationMSMInputsForLtmle function.}

\item{summary.measures}{Summary measures as output by the GetMediationMSMInputsForLtmle function.}

\item{final.Ynodes}{Final Y node(s).}

\item{stratify}{Logical enabling stratified outcome.}

\item{msm.weights}{MSM weights.}

\item{estimate.time}{Measure time to fun function.}

\item{gcomp}{gcomp formula.}

\item{iptw.only}{Use IPTW estimator only}

\item{deterministic.Q.function}{Logical specifying if Q is a deterministic function.}

\item{IC.variance.only}{Only estimate variance through the influence curve}

\item{observation.weights}{Provide weight for observations}
}
\value{
Returns output ready for ltmleMed.
}
\description{
Create Mediation Inputs for ltmle.
}
