% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medltmle_CreateMedMSMInputs.R
\name{CreateMedMSMInputs}
\alias{CreateMedMSMInputs}
\title{CreateMedMSMInputs}
\usage{
CreateMedMSMInputs(data, abar, abar.prime, rule, gform)
}
\arguments{
\item{data}{Dataframe containing the data in a wide format.}

\item{abar}{Intervention.}

\item{abar.prime}{Control for the exposure.}

\item{rule}{Specify rule for the intervention.}

\item{gform}{g form for intervention (if there is a censoring variable, include C as well).}
}
\value{
Returns regimes for abar and abar.prime in proper format, working MSM and MSM weights, summary measures, gform and final Y nodes.
}
\description{
Create Mediation MSM Inputs for ltmle.
}
