% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medltmle_HelperFun.R
\name{CreateNodes}
\alias{CreateNodes}
\title{CreateNodes}
\usage{
CreateNodes(data, Anodes, Cnodes, Lnodes, Ynodes, Znodes = NULL)
}
\arguments{
\item{data}{TO DO}

\item{Anodes}{TO DO}

\item{Cnodes}{TO DO}

\item{Lnodes}{TO DO}

\item{Ynodes}{TO DO}

\item{Znodes}{TO DO}
}
\value{
Returns all nodes.
}
\description{
Get all the nodes from input data.
}
