% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medltmle_HelperFun.R
\name{dataConvert}
\alias{dataConvert}
\title{dataConvert}
\usage{
dataConvert(data, nBCov, type)
}
\arguments{
\item{data}{Dataframe object in a wide format.}

\item{nBCov}{number of baseline covariates. They must the listed before C,A,Z,LA,LZ,Y,D nodes.}

\item{type}{Which format to convert the current data to, long or wide.}
}
\value{
Dataframe object of the original data in a long or wide format.
}
\description{
Function to convert wide to long, or long to wide dataframe.
}
