% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltmleMediation_TMLECalcs.R
\name{Estimate}
\alias{Estimate}
\title{Estimate}
\usage{
Estimate(inputs, form, subs, family, type, nodes, Qstar.kplus1, cur.node,
  calc.meanL, called.from.estimate.g, regimes.meanL,
  regimes.with.positive.weight)
}
\arguments{
\item{inputs}{Output of \code{CreateMedInputs}}

\item{form}{Q form for current node.}

\item{subs}{Subset of samples for stratify option.}

\item{family}{TO DO}

\item{type}{Two options, "response" or "link".}

\item{nodes}{TO DO}

\item{Qstar.kplus1}{TO DO}

\item{cur.node}{TO DO}

\item{calc.meanL}{Defaults to \code{!IC.variance.only}.}

\item{called.from.estimate.g}{TO DO}

\item{regimes.meanL}{TO DO}

\item{regimes.with.positive.weight}{Defaults to \code{1:num.regimes}.}
}
\value{
Returns predicted values for the fit.
}
\description{
Run GLM or SuperLearner to obtain an estimate for the current node.
}
