% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltmleMediation_TMLECalcs.R
\name{EstimateG}
\alias{EstimateG}
\title{EstimateG}
\usage{
EstimateG(inputs, regimes.use)
}
\arguments{
\item{inputs}{Output of \code{CreateMedInputs}}

\item{regimes.use}{Indicate which regime to follow.}
}
\value{
Returns estimate of conditional density for each A node, bounded and unbounded cumulative g.
}
\description{
Estimation of conditional densities of A nodes (g)
}
