% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltmleMediation_TMLECalcs.R
\name{EstimateMultiDens}
\alias{EstimateMultiDens}
\title{EstimateMultiDens}
\usage{
EstimateMultiDens(inputs, use.regimes, use.intervention.match, is.Z.dens)
}
\arguments{
\item{inputs}{Output of \code{CreateMedInputs}}

\item{use.regimes}{TO DO}

\item{use.intervention.match}{TO DO}

\item{is.Z.dens}{TO DO}
}
\value{
Returns estimate of L densities.
}
\description{
If not is.Z.dens, estimating L densities
}
