% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltmleMediation_TMLECalcs.R
\name{FinalizeIC}
\alias{FinalizeIC}
\title{FinalizeIC}
\usage{
FinalizeIC(
  IC,
  combined.summary.measures,
  Qstar,
  m.beta,
  msm.weights,
  observation.weights
)
}
\arguments{
\item{IC}{TO DO}

\item{combined.summary.measures}{TO DO}

\item{Qstar}{TO DO}

\item{m.beta}{TO DO}

\item{msm.weights}{TO DO}

\item{observation.weights}{TO DO}
}
\value{
Returns final Influence Curve.
}
\description{
Final step in calculating TMLE influence curve
}
