% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltmleMediation_TMLECalcs.R
\name{FixedTimeTMLEMediation}
\alias{FixedTimeTMLEMediation}
\title{FixedTimeTMLEMediation}
\usage{
FixedTimeTMLEMediation(inputs, nodes, msm.weights, combined.summary.measures,
  g.abar.list, g.abar.prime.list, cum.qz.abar, cum.qz.abar.prime, cum.qL.abar,
  cum.qL.abar.prime)
}
\arguments{
\item{inputs}{Output of \code{CreateMedInputs}}

\item{nodes}{All nodes in the data.}

\item{msm.weights}{Marginal structural model weights.}

\item{combined.summary.measures}{TO DO}

\item{g.abar.list}{g estimate from \code{EstimateG} with regime abar.}

\item{g.abar.prime.list}{g estimate from \code{EstimateG} with regime abar.prime.}

\item{cum.qz.abar}{Cumulative estimate for Z nodes under abar regime.}

\item{cum.qz.abar.prime}{Cumulative estimate for Z nodes under abar.prime regime.}

\item{cum.qL.abar}{Cumulative estimate for L nodes under abar regime.}

\item{cum.qL.abar.prime}{Cumulative estimate for L nodes under abar.prime regime.}
}
\value{
Return...
}
\description{
TMLE Mediation
}
