% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltmleMediation_TMLECalcs.R
\name{FixScoreEquation}
\alias{FixScoreEquation}
\title{FixScoreEquation}
\usage{
FixScoreEquation(
  Qstar.kplus1,
  h.g.ratio,
  uncensored,
  intervention.match,
  is.deterministic,
  deterministic.Q,
  off,
  X,
  regimes.with.positive.weight
)
}
\arguments{
\item{Qstar.kplus1}{Estimate of the expectation with respect to the distribution of the node one ahead of the current node given the past (dimension: n x num.regimes).}

\item{h.g.ratio}{I() * h * observation.weights * cum.q.ratio / g}

\item{uncensored}{Uncensored samples}

\item{intervention.match}{Samples that match the rule.}

\item{is.deterministic}{TO DO}

\item{deterministic.Q}{TO DO}

\item{off}{offset used for TMLE.}

\item{X}{TO DO}

\item{regimes.with.positive.weight}{Regimes with positive weight.}
}
\value{
Returns solved score equation.
}
\description{
In case the GLM did not converge and the TMLE does not solve the score equation, attempt to solve the score equation directly.
}
