% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateData.R
\name{GenerateData}
\alias{GenerateData}
\title{GenerateData}
\usage{
GenerateData(n, end.time, abar = NULL, abar.prime = NULL)
}
\arguments{
\item{n}{Sample size}

\item{end.time}{Last time point}

\item{abar}{Set intervention}

\item{abar.prime}{Control intervention}
}
\value{
Returns a dataframe with simulated covariates as described in the description.
}
\description{
Simulate longitudinal data as specified in the description. Note that the data is of the following format:
O={W1 W2 C1 A1 LA1 Z1 LZ1 Y1 C2 A2 LA2 Z2 LZ2 Y2 ... }
}
