% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medltmle_CreateMedInputs.R
\name{GetDefaultFormMediation}
\alias{GetDefaultFormMediation}
\title{GetDefaultFormMediation}
\usage{
GetDefaultFormMediation(data, nodes, is.Qform, is.QLform, is.qzform, is.qLform,
  past, time.end, stratify, survivalOutcome, YisL = YisL, showMessage)
}
\arguments{
\item{data}{Available data in a \code{Data.Frame} format.}

\item{nodes}{List of available nodes, as created by \code{CreateNodes}.}

\item{is.Qform}{Logical indicating whether to specify general Q formula.}

\item{is.QLform}{Logical indicating whether to specify Q formula for covariates.}

\item{is.qzform}{Logical indicating whether to specify general Q formula for Z.}

\item{is.qLform}{Logical indicating whether to specify general Q formula for L.}

\item{past}{Number indicating Markov order for the conditional densities.}

\item{time.end}{Total number of time points.}

\item{stratify}{Logical indicating whether to straify.}

\item{survivalOutcome}{Logical indicating if the outcome a survival outcome.}

\item{YisL}{Logical indicating whether Y is a function of time-varying covariates.}

\item{showMessage}{Logical indicating whether to show comments while executing.}
}
\value{
Returns default Q or g formula if not specified.
}
\description{
If QLform, QZform, qzform and gform are not specified, return default form.
Each formula consists of all parent nodes except censoring and event nodes.
If \code{stratify}=TRUE, do not include A nodes.
}
