% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medltmle_HelperFun.R
\name{GetLibrary}
\alias{GetLibrary}
\title{GetLibrary}
\usage{
GetLibrary(SL.library, estimate.type)
}
\arguments{
\item{SL.library}{TO DO}

\item{estimate.type}{TO DO}
}
\value{
Returns Super Learner library.
}
\description{
Get Super Learner library by type (Q or g).
}
