% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMediationMSMInputsForLtmle.R
\name{GetMediationMSMInputsForLtmle}
\alias{GetMediationMSMInputsForLtmle}
\title{GetMediationMSMInputsForLtmle}
\usage{
GetMediationMSMInputsForLtmle(data, abar, abar.prime, rule, gform)
}
\arguments{
\item{data}{Dataframe containing the data in a wide format.}

\item{abar}{Intervention.}

\item{abar.prime}{Control for the exposure.}

\item{rule}{Specify rule for the intervention.}

\item{gform}{g form for intervention (if there is a censoring variable, include C as well).}
}
\value{
Returns regimes for abar and abar.prime, working MSM and MSM weights, summary measures, gform and final Y nodes.
}
\description{
Get Mediation MSM Inputs for ltmle.
}
