% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltmleMediation_TMLECalcs.R
\name{GetMsmWeights}
\alias{GetMsmWeights}
\title{GetMsmWeights}
\usage{
GetMsmWeights(inputs)
}
\arguments{
\item{inputs}{Output of \code{CreateMedInputs}.}
}
\value{
Returns weights for the Marginal Structural Model.
}
\description{
Set weights for the Marginal Structural Model
}
