% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medltmle_HelperFun.R
\name{InterventionMatch}
\alias{InterventionMatch}
\title{InterventionMatch}
\usage{
InterventionMatch(intervention.match.array, Anodes, cur.node)
}
\arguments{
\item{intervention.match.array}{TO DO}

\item{Anodes}{TO DO}

\item{cur.node}{TO DO}
}
\value{
Returns matrix of [numObservations x numRegimes] I(A==abar) from Anodes[1] to the Anode just before cur.node.
}
\description{
Determine which patients are following specified treatment regime.
}
