% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medltmle_HelperFun.R
\name{IsDeterministic}
\alias{IsDeterministic}
\title{IsDeterministic}
\usage{
IsDeterministic(
  data,
  cur.node,
  deterministic.Q.function,
  nodes,
  called.from.estimate.g,
  survivalOutcome
)
}
\arguments{
\item{data}{TO DO}

\item{cur.node}{TO DO}

\item{deterministic.Q.function}{TO DO}

\item{nodes}{TO DO}

\item{called.from.estimate.g}{TO DO}

\item{survivalOutcome}{TO DO}
}
\value{
Return list contains:
is.deterministic: vector of [numObservations x 1] - true if patient is already dead before cur.node or set by deterministic.Q.function
Q.value: vector of [which(is.deterministic) x 1] - value of Q
}
\description{
Determine which patients have died or have Q set deterministically by user function before cur.node.
}
