% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medltmle_HelperFun.R
\name{IsDeterministicG}
\alias{IsDeterministicG}
\title{IsDeterministicG}
\usage{
IsDeterministicG(data, cur.node, deterministic.g.function, nodes, using.newdata)
}
\arguments{
\item{data}{TO DO}

\item{cur.node}{TO DO}

\item{deterministic.g.function}{TO DO}

\item{nodes}{TO DO}

\item{using.newdata}{TO DO}
}
\value{
Returns TRUE for patients that have a deterministic Anode value.
}
\description{
Determines which patients have an Anode value which is deterministic. For example, deterministic.g.function may be used to
specify that once a patient starts treatment, they stay on treatment and this should be taken into
consideration during estimation of G.
}
