% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medltmle_ltmleMediation.R
\name{ltmleMediation}
\alias{ltmleMediation}
\title{ltmleMediation}
\usage{
ltmleMediation(inputs)
}
\arguments{
\item{inputs}{Output of \code{CreateMedInputs}}
}
\value{
Returns estimate of $E[Y_{\tau}(a, \overline{\Gamma}^{a^'})]$
}
\description{
Longitudinal Mediation Analysis with Time-varying Mediators.
}
