% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medltmle_ltmleMediation.R
\name{LtmleMediationMSMFromInputs}
\alias{LtmleMediationMSMFromInputs}
\title{LtmleMediationMSMFromInputs}
\usage{
LtmleMediationMSMFromInputs(inputs)
}
\arguments{
\item{inputs}{Output of \code{CreateMedInputs}.}
}
\value{
Returns results from \code{MainCalcsMediation}, and adds Gcomp, formulas,
biinary outcome, transformed outcome and survival outcome logical to the results from \code{CreateMedInputs}.
}
\description{
Calls \code{MainCalcsMediation} and edits the results obtained.
}
