% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medltmle.R
\name{medltmle}
\alias{medltmle}
\title{medltmle}
\usage{
medltmle(data, Anodes, Znodes, Cnodes = NULL, Lnodes = NULL, Ynodes,
  W2nodes = NULL, Dnodes = NULL, survivalOutcome = NULL, QLform = NULL,
  QZform = NULL, gform = NULL, qzform = NULL, qLform = NULL, abar,
  abar.prime, rule = NULL, gbounds = c(0.01, 1), Yrange = NULL,
  deterministic.g.function = NULL, deterministic.Q.function = NULL,
  stratify = FALSE, SL.library = NULL, estimate.time = TRUE,
  gcomp = FALSE, iptw.only = FALSE, IC.variance.only = FALSE,
  observation.weights = NULL, CSE, time.end, past, YisL = TRUE)
}
\arguments{
\item{data}{Dataframe containing the data in a wide format.}

\item{Anodes}{names of columns containing A covariates (exposure) (character).}

\item{Znodes}{names of columns containing Z covariates (mediator) (character).}

\item{Cnodes}{names of columns containing C covariates (censoring) (character).}

\item{Lnodes}{names of columns containing L covariates (covariate) (character).}

\item{Ynodes}{names of columns containing Y covariates (outcome) (character).}

\item{W2nodes}{names of columns containing W2 covariates (baseline covariates in need of fluctuation) (character).}

\item{Dnodes}{names of columns containing D covariates (death indicator) (character).}

\item{survivalOutcome}{logical variable specifying if the outcome is survival.}

\item{QLform}{Q forms for L covariates.}

\item{QZform}{Q forms for Z covariates.}

\item{gform}{g form for intervention (if there is a censoring variable, include C as well).}

\item{qzform}{g form for Z covariates.}

\item{qLform}{g form for L covariates.}

\item{abar}{Intervention. Dimension should correspond to the number of intervention nodes available.}

\item{abar.prime}{Control for the exposure. Dimension should correspond to the number of intervention nodes available.}

\item{rule}{Specify rule for the intervention.}

\item{gbounds}{Bounds for the propensity score.}

\item{Yrange}{Specify range for the outcome.}

\item{deterministic.g.function}{Logical specifying if g is a deterministic function.}

\item{deterministic.Q.function}{Logical specifying if Q is a deterministic function.}

\item{stratify}{Logical enabling stratified outcome.}

\item{SL.library}{SuperLearner library for estimation.}

\item{estimate.time}{Measure time to fun function.}

\item{gcomp}{Logical indicating whether to use Gcomp instead (no updating if TRUE).}

\item{iptw.only}{Use IPTW estimator only}

\item{IC.variance.only}{Only estimate variance through the influence curve}

\item{observation.weights}{Provide weight for observations}

\item{CSE}{Logical specifying if the estimand is estimated by fully conditioning on the past (TRUE), or with the data-dependent estimate (FALSE).}

\item{time.end}{How many time points in the longitudinal data?}

\item{YisL}{Logical indicating whether Y is a function of time-varying covariate.}
}
\value{
Returns estimate of \eqn{E[Y_{\tau}(a, \overline{\Gamma}^{a^'})]}
}
\description{
Estimates parameters for longitudinal mediation analysis with time-varying mediators.
}
