% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medltmle_HelperFun.R
\name{NodeToIndex}
\alias{NodeToIndex}
\title{NodeToIndex}
\usage{
NodeToIndex(data, node)
}
\arguments{
\item{data}{TO DO}

\item{node}{TO DO}
}
\value{
Returns index of the nodes.
}
\description{
Convert named nodes to indicies of nodes
}
