% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltmleMediation_TMLECalcs.R
\name{NormalizeIC}
\alias{NormalizeIC}
\title{NormalizeIC}
\usage{
NormalizeIC(IC, combined.summary.measures, m.beta, msm.weights,
  observation.weights, g.ratio)
}
\arguments{
\item{IC}{TO DO}

\item{combined.summary.measures}{TO DO}

\item{m.beta}{TO DO}

\item{msm.weights}{TO DO}

\item{observation.weights}{TO DO}

\item{g.ratio}{TO DO}
}
\value{
Returns
}
\description{
Normalize the influence curve matrix
}
