% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medltmle_HelperFun.R
\name{RegimesFromAbar}
\alias{RegimesFromAbar}
\title{RegimesFromAbar}
\usage{
RegimesFromAbar(data, abar, rule)
}
\arguments{
\item{data}{TO DO}

\item{abar}{TO DO}

\item{rule}{TO DO}
}
\value{
Returns regimes.
}
\description{
Creates regimes from abar and abar.prime
}
