% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medltmle_HelperFun.R
\name{RhsVars}
\alias{RhsVars}
\title{RhsVars}
\usage{
RhsVars(f)
}
\arguments{
\item{f}{TO DO}
}
\value{
Return the right hand side variables of formula f as a character vector
}
\description{
Returns the right hand side variables of a formula.
}
