% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medltmle_HelperFun.R
\name{SetA}
\alias{SetA}
\title{SetA}
\usage{
SetA(data, regimes, Anodes, regime.index, cur.node)
}
\arguments{
\item{data}{TO DO}

\item{regimes}{TO DO}

\item{Anodes}{TO DO}

\item{regime.index}{TO DO}

\item{cur.node}{TO DO}
}
\value{
Returns data with Anodes set to regime.
}
\description{
Set the Anodes of data to regime[, , regime.index] up to cur.node
}
