% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medltmle_HelperFun.R
\name{SetALA}
\alias{SetALA}
\title{SetALA}
\usage{
SetALA(data, regimes, Anodes, regime.index, cur.node, regimes_add)
}
\arguments{
\item{data}{\code{data.frame} object containing the data.}

\item{regimes}{which regime should the intervention nodes follow.}

\item{Anodes}{column indexes for A nodes in the data.}

\item{regime.index}{index of the regime set to follow.}

\item{cur.node}{current node}

\item{regimes_add}{additional regime set for a node after the mediator.}
}
\value{
Returns data with Anodes and LAnodes set to regime.
}
\description{
Set the Anodes of data to regime[, , regime.index] up to cur.node, and the closet LA node to specified values.
}
