% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medltmle_HelperFun.R
\name{timeOrder_baseline}
\alias{timeOrder_baseline}
\title{timeOrder_baseline}
\usage{
timeOrder_baseline(data, A)
}
\arguments{
\item{data}{data.frame object containing all the baseline covariates and possibly an exposure.}

\item{A}{Numeric value indicating the column number of exposure. Must be within (1,ncol(data)) range.}
}
\value{
Returns data with set exposure A and ordered baseline covariates. All baseline covariates before A
are W1, whereas all after A are W2 covariates. This part is necessary later for TMLE fluctuation.
}
\description{
Function to group baseline covariates according to time ordering, in case there is time dependence.
}
