% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltmleMediation_TMLECalcs.R
\name{UpdateQMediation}
\alias{UpdateQMediation}
\title{UpdateQMediation}
\usage{
UpdateQMediation(Qstar.kplus1, logitQ, combined.summary.measures, cum.g,
  cum.q.ratio, working.msm, uncensored, intervention.match, is.deterministic,
  msm.weights, gcomp, observation.weights)
}
\arguments{
\item{Qstar.kplus1}{TO DO}

\item{logitQ}{TO DO}

\item{combined.summary.measures}{TO DO}

\item{cum.g}{TO DO}

\item{cum.q.ratio}{TO DO}

\item{working.msm}{TO DO}

\item{uncensored}{TO DO}

\item{intervention.match}{TO DO}

\item{is.deterministic}{TO DO}

\item{msm.weights}{TO DO}

\item{gcomp}{TO DO}

\item{observation.weights}{TO DO}
}
\value{
Returns targeted Q.
}
\description{
Update the initial fit of Q using clever covariates.
}
