% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltmleMediation_TMLECalcs.R
\name{CalcCumG}
\alias{CalcCumG}
\title{CalcCumG}
\usage{
CalcCumG(g, gbounds)
}
\arguments{
\item{g}{TO DO}

\item{gbounds}{TO DO}
}
\value{
Returns G.
}
\description{
Calculate bounded cumulative G
}
