% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltmleMediation_TMLECalcs.R
\name{CalcG}
\alias{CalcG}
\title{CalcG}
\usage{
CalcG(prob.A.is.1, cur.abar, deterministic.newdata)
}
\arguments{
\item{prob.A.is.1}{Probability A is 1.}

\item{cur.abar}{Regime for the exposure.}

\item{deterministic.newdata}{Logical for whether samples are deterministic.}
}
\value{
Returns G.
}
\description{
Calculate G
}
