% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltmleMediation_TMLECalcs.R
\name{CalcIC}
\alias{CalcIC}
\title{CalcIC}
\usage{
CalcIC(Qstar.kplus1, Qstar, h.g.ratio, uncensored, intervention.match,
  regimes.with.positive.weight)
}
\arguments{
\item{Qstar.kplus1}{TO DO}

\item{Qstar}{TO DO}

\item{h.g.ratio}{TO DO}

\item{uncensored}{TO DO}

\item{intervention.match}{TO DO}

\item{regimes.with.positive.weight}{TO DO}
}
\value{
Returns IC for one node.
}
\description{
Calculate the TMLE influence curve for one node.
}
