% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medltmle_HelperFun.R
\name{CalcInterventionMatchArray}
\alias{CalcInterventionMatchArray}
\title{CalcInterventionMatchArray}
\usage{
CalcInterventionMatchArray(data, regimes, Anodes)
}
\arguments{
\item{data}{TO DO}

\item{regimes}{TO DO}

\item{Anodes}{Index of A nodes. Can be easily derived from \code{all.nodes$A}.}
}
\value{
Returns a logical array, n x num.Anodes x num.regimes = follows regime[j] up to and including Anode[i]. NA will count as a match as well.
}
\description{
Check if A as in data equals the regime.
}
