% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltmleMediation_TMLECalcs.R
\name{CalcIPTWMediation}
\alias{CalcIPTWMediation}
\title{CalcIPTWMediation}
\usage{
CalcIPTWMediation(inputs, cum.g.abar, cum.qz.abar, cum.qz.abar.prime,
  msm.weights, estimand)
}
\arguments{
\item{inputs}{Output of \code{CreateMedInputs}}

\item{cum.g.abar}{Cumulative g obtained from \code{CalcCumG}  with regime abar.}

\item{cum.qz.abar}{Cumulative Q for Z obtained from \code{CalcCumG} with regime abar.}

\item{cum.qz.abar.prime}{Cumulative Q for Z obtained from \code{CalcCumG} with regime abar.prime.}

\item{msm.weights}{Marginal Structural Model weights for each sample.}

\item{estimand}{Specifies which estimand to estimate. Options are: natural effect (NE), stochastic effect (SE), or controlled effect (CE).}
}
\value{
Returns IPTW estimate and normalized influence curve for IPTW.
}
\description{
Calculates IPTW estimate.
}
