% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medltmle_HelperFun.R
\name{CreateNodes}
\alias{CreateNodes}
\title{CreateNodes}
\usage{
CreateNodes(
  data,
  Anodes,
  Cnodes,
  Lnodes,
  Ynodes,
  Inodes = NULL,
  Znodes = NULL,
  Dnodes = NULL,
  W2nodes = NULL
)
}
\arguments{
\item{data}{Available data in a wide format.}

\item{Anodes}{Character variable with exposure nodes.}

\item{Cnodes}{Character variable with censoring nodes.}

\item{Lnodes}{Character variable with covariate nodes.}

\item{Ynodes}{Character variable with outcome nodes.}

\item{Inodes}{Character variable with instrumental nodes.}

\item{Znodes}{Character variable with mediator nodes.}

\item{Dnodes}{Character variable with death as censoring nodes.}

\item{W2nodes}{Character variable with baseline nodes after treatment that need to be fluctuated during the TMLE procedure.}
}
\value{
Returns all nodes in a format necessary for further functions.
}
\description{
Get all the nodes from input data.
}
