% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltmleMediation_TMLECalcs.R
\name{Estimate}
\alias{Estimate}
\title{Estimate}
\usage{
Estimate(
  inputs,
  form,
  subs,
  family,
  type,
  nodes,
  Qstar.kplus1,
  cur.node,
  calc.meanL,
  called.from.estimate.g,
  regimes.meanL,
  regimes.with.positive.weight,
  CSE_I = FALSE,
  CSE_Z = FALSE,
  regimes_add = NULL
)
}
\arguments{
\item{inputs}{Output of \code{CreateMedInputs}.}

\item{form}{Q form for current node.}

\item{subs}{Logical indicating uncensored and non-deterministic samples.}

\item{family}{a description of the error distribution and link function to be used in the model.}

\item{type}{"response" or "link"}

\item{nodes}{List with index for each node subset. Output of \code{CreateNodes}.}

\item{Qstar.kplus1}{Estimate of the expectation with respect to the distribution of the node one ahead of the current node given the past (dimension: n x num.regimes).}

\item{cur.node}{Node being estimated.}

\item{calc.meanL}{Estimate conditional density of A using mean of covariates. This is a useful option if there are NAs in the regime, or want to estimate variance.}

\item{called.from.estimate.g}{Logical TO DO}

\item{regimes.meanL}{TO DO}

\item{regimes.with.positive.weight}{Regimes with positve weights. Defaults to \code{1:num.regimes}.}

\item{CSE_I}{Logical indicating whether this is estimation with instrumental variable.}

\item{CSE_Z}{Logical indicating whether this is a Z estimation for the data-dependent parameter.}

\item{regimes_add}{If CSE_Z is TRUE, must specify at what value the node after the mediator should be estimated at.}
}
\value{
Returns predicted values for the fit as well as the fit as well as indicators for which samples are deterministic and their values.
If specified, it also returns the probability of A=1 given mean L.
}
\description{
Run GLM or SuperLearner to obtain an estimate for the current node.
}
