% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltmleMediation_TMLECalcs.R
\name{EstimateG}
\alias{EstimateG}
\title{EstimateG}
\usage{
EstimateG(inputs, regimes.use)
}
\arguments{
\item{inputs}{Output of \code{CreateMedInputs}}

\item{regimes.use}{}
}
\value{
Returns
}
\description{
Parametric estimation of each g-factor.
}
