% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltmleMediation_TMLECalcs.R
\name{EstimateMultiDens}
\alias{EstimateMultiDens}
\title{EstimateMultiDens}
\usage{
EstimateMultiDens(inputs, use.regimes, use.intervention.match, is.Z.dens,
  estimand)
}
\arguments{
\item{inputs}{Output of \code{CreateMedInputs}}

\item{use.regimes}{TO DO}

\item{use.intervention.match}{TO DO}

\item{is.Z.dens}{Indicates whether the conditional density being estimated is for the Z node.}

\item{estimand}{Specifies which estimand to estimate. Options are: natural effect (NE), stochastic effect (SE), or controlled effect (CE).}
}
\value{
Returns estimate of Z and L conditional densities.
}
\description{
Estimates conditional densities of Z and L.
}
