% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medltmle_CreateMedInputs.R
\name{GetDefaultFormMediation}
\alias{GetDefaultFormMediation}
\title{GetDefaultFormMediation}
\usage{
GetDefaultFormMediation(data, nodes, is.Qform, is.QLform, is.qzform, stratify,
  survivalOutcome, showMessage)
}
\arguments{
\item{data}{TO DO}

\item{nodes}{TO DO}

\item{is.Qform}{TO DO}

\item{is.QLform}{TO DO}

\item{is.qzform}{TO DO}

\item{stratify}{TO DO}

\item{survivalOutcome}{TO DO}

\item{showMessage}{TO DO}
}
\value{
Returns default Q or g formula.
}
\description{
Get the default Q or g formula, if not specified.
Each formula consists of all parent nodes except censoring and event nodes.
If \code{stratify}=TRUE, do not include A nodes.
}
