% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medltmle_HelperFun.R
\name{GetLibrary}
\alias{GetLibrary}
\title{GetLibrary}
\usage{
GetLibrary(SL.library, estimate.type)
}
\arguments{
\item{SL.library}{List containing two elements: Super Learner algorithms for Q estimation, and Super Learner algorithms for g estimation.}

\item{estimate.type}{Q or g library.}
}
\value{
Returns Super Learner library for Q or g.
}
\description{
Get Super Learner library by type (Q or g).
}
