% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medltmle_HelperFun.R
\name{IsUncensored}
\alias{IsUncensored}
\title{IsUncensored}
\usage{
IsUncensored(uncensored.matrix, Cnodes, cur.node)
}
\arguments{
\item{uncensored.matrix}{TO DO}

\item{Cnodes}{TO DO}

\item{cur.node}{TO DO}
}
\value{
Returns vector of [numDataRows x 1] I(C=uncensored) from Cnodes[1] to the Cnode just before cur.node
}
\description{
Determine which patients are uncensored
}
