% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medltmle_ltmleMediation.R
\name{ltmleMediation}
\alias{ltmleMediation}
\title{ltmleMediation}
\usage{
ltmleMediation(inputs)
}
\arguments{
\item{inputs}{Output of \code{CreateMedInputs}}
}
\value{
Returns $psi^{IPW}_n$, $psi^{TMLE}_n$ and if specified $psi^{NTsub}_n$ estimates,
corresponding influence curve, cumulative conditional density with abar and abar.prime rule for g, QL and QZ (bounded and unbounded),
fits for Q and Qstar and corresponding formulas, final Qstar estimates, logical indicating binary and/or transformed outcome.
}
\description{
Prepares output obtained from \code{LtmleMediationMSMFromInputs} and \code{MainCalcsMediation} for final report by
\code{medltmle}.
}
