% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medltmle_ltmleMediation.R
\name{LtmleMediationMSMFromInputs}
\alias{LtmleMediationMSMFromInputs}
\title{LtmleMediationMSMFromInputs}
\usage{
LtmleMediationMSMFromInputs(inputs)
}
\arguments{
\item{inputs}{Output of \code{CreateMedInputs}}
}
\value{
Returns
}
\description{
run ltmleMSM from ltmleInputs object
}
