% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltmleMediation_TMLECalcs.R
\name{MainCalcsMediation}
\alias{MainCalcsMediation}
\title{MainCalcsMediation}
\usage{
MainCalcsMediation(inputs)
}
\arguments{
\item{inputs}{Output of \code{CreateMedInputs}}
}
\value{
Returns the influence curve for TMLE and IPTW estimators, marginal structural model fit, cumulative conditional density estimates for Qs and g,
separate fits for each node, variance estimate if specified, IPTW estimate and final Qstar.
}
\description{
Main TMLE calculations.
}
