% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltmleMediation_TMLECalcs.R
\name{MainCalcsMediation}
\alias{MainCalcsMediation}
\title{MainCalcsMediation}
\usage{
MainCalcsMediation(inputs)
}
\arguments{
\item{inputs}{Output of \code{CreateMedInputs}}
}
\value{
Returns
}
\description{
Main TMLE calculations.
}
