% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medltmle.R
\name{medltmle}
\alias{medltmle}
\title{medltmle}
\usage{
medltmle(data, Anodes, Znodes, Cnodes = NULL, Lnodes = NULL, Ynodes,
  survivalOutcome = NULL, QLform = NULL, QZform = NULL, gform = NULL,
  qzform = NULL, qLform = NULL, abar, abar.prime, rule = NULL,
  gbounds = c(0.01, 1), Yrange = NULL, deterministic.g.function = NULL,
  deterministic.Q.function = NULL, stratify = FALSE, SL.library = NULL,
  estimate.time = TRUE, gcomp = FALSE, iptw.only = FALSE,
  IC.variance.only = FALSE, observation.weights = NULL)
}
\arguments{
\item{data}{Dataframe containing the data in a wide format.}

\item{Anodes}{names of columns containing A covariates (character).}

\item{Znodes}{names of columns containing Z covariates (character).}

\item{Cnodes}{names of columns containing C covariates (character).}

\item{Lnodes}{names of columns containing L covariates (character).}

\item{Ynodes}{names of columns containing Y covariates (character).}

\item{survivalOutcome}{logical variable specifying if the outcome is survival.}

\item{QLform}{Q forms for L covariates.}

\item{QZform}{Q forms for Z covariates.}

\item{gform}{g form for intervention (if there is a censoring variable, include C as well).}

\item{qzform}{g form for Z covariates.}

\item{qLform}{g form for L covariates.}

\item{abar}{Intervention. Dimension should correspond to the number of intervention nodes available.}

\item{abar.prime}{Control for the exposure. Dimension should correspond to the number of intervention nodes available.}

\item{rule}{Specify rule for the intervention.}

\item{gbounds}{Bounds for the propensity score.}

\item{Yrange}{Specify range for the outcome.}

\item{deterministic.g.function}{Logical specifying if g is a deterministic function.}

\item{deterministic.Q.function}{Logical specifying if Q is a deterministic function.}

\item{stratify}{Logical enabling stratified outcome.}

\item{SL.library}{SuperLearner library for estimation.}

\item{estimate.time}{Measure time to fun function.}

\item{gcomp}{Logical indicating whether to use Gcomp instead (no updating if TRUE).}

\item{iptw.only}{Use IPTW estimator only}

\item{IC.variance.only}{Only estimate variance through the influence curve}

\item{observation.weights}{Provide weight for observations}
}
\value{
Returns estimate of $E[Y_{\tau}(a, \overline{\Gamma}^{a^'})]$
}
\description{
Estimates parameters for longitudinal mediation analysis with time-varying mediators.
}
