% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltmleMediation_TMLECalcs.R
\name{NormalizeIC}
\alias{NormalizeIC}
\title{NormalizeIC}
\usage{
NormalizeIC(IC, combined.summary.measures, m.beta, msm.weights,
  observation.weights, g.ratio)
}
\arguments{
\item{IC}{Estimate of the influence curve.}

\item{combined.summary.measures}{Combined summary measure of (n x num.measures x num.regimes x num.final.Ynodes) dimension. Note that num.measures is equal to num.summary.measures + num.baseline.covariates.}

\item{m.beta}{Estimate of betas.}

\item{msm.weights}{Marginal structural model weights.}

\item{observation.weights}{Observation weights.}

\item{g.ratio}{\code{g.unbounded} / \code{g.bounded} ratio of dimensions: n x num.regimes x num.final.Ynodes. If \code{IC.variance.only}, g.ratio should be NULL.}
}
\value{
Returns normalized influence curve.
}
\description{
Normalize the influence curve matrix.
}
