% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegimesFromAbar.R
\name{RegimesFromAbar}
\alias{RegimesFromAbar}
\title{RegimesFromAbar}
\usage{
RegimesFromAbar(data, abar, rule)
}
\arguments{
\item{data}{Dataframe containing the data in a wide format.}

\item{abar}{Intervention.}

\item{rule}{Specify rule for the intervention.}
}
\value{
Returns regimes for the interventions.
}
\description{
Helper function for GetMediationMSMInputsForLtmle.
}
