% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medltmle_HelperFun.R
\name{TransformOutcomes}
\alias{TransformOutcomes}
\title{TransformOutcomes}
\usage{
TransformOutcomes(data, nodes, Yrange)
}
\arguments{
\item{data}{TO DO}

\item{nodes}{TO DO}

\item{Yrange}{TO DO}
}
\value{
Returns data with transfored outcome.
}
\description{
TransformOutcomes
}
